Change sortlist prototype to take const void **.
This is to make it clearer that this is what the compare function needs to be.
It needs to be this because the list contains pointers, and "qsort" will call
the compare function with a pointer to the list element, i.e. a pointer to a
pointer.  "qsort" doesn't know the list contains pointers, which is why we
have to cast the function to const void * before passing it to "qsort".