Round block count up if BLOCKSIZE > blocksize.
This is to match the behavior of BSD ls.  Logically, this is what we
want anyway.  It tells us how many blocks would be needed if blocks were
BLOCKSIZE bytes.

For example, if stat tells us the file uses 3 512-byte blocks,
and therefore has a size of <= 1536 bytes, it would use 2 1024-byte
blocks, not 1.